att_calculadora

+++ src/test/java/br/gov/sp/fatec/projetomaven/calculadora.java
@@ -10,11 +10,11 @@ public class calculadora {
 		return n1 - n2;
 	}
 	
-	public int multiplicacao(int n1, int n2) {	
+	public double multiplicacao(double n1, double n2) {	
 		return n1 * n2;
 	}
 	
-	public int divisao(int n1, int n2) {
+	public double divisao(double n1, double n2) {
 		return n1 / n2;
 	}
 }
